backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nuevo-proyecto-backend"
version = "0.1.0"
description = "Backend base con arquitectura hexagonal"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115.0,<1.0.0",
  "uvicorn[standard]>=0.30.0,<1.0.0",
  "sqlalchemy>=2.0.0,<3.0.0",
  "alembic>=1.13.0,<2.0.0",
  "pydantic>=2.8.0,<3.0.0",
  "google-auth>=2.27.0,<3.0.0",
  "google-auth-oauthlib>=1.1.0,<2.0.0",
  "google-api-python-client>=2.0.0,<3.0.0",
  "pydantic-extra-types>=2.0.0,<3.0.0",
  "python-multipart>=0.0.7,<0.1.0",
  "httpx>=0.26.0,<1.0.0",
  "orjson>=3.9.0,<4.0.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "pytest-cov>=5.0.0",
  "pytest-env>=1.1.0",
  "httpx>=0.27.0",
  "ruff>=0.4.0",
  "import-linter>=2.0",
  "pytest-asyncio>=0.24.0",
  "google-auth-stubs>=0.1.0",
  "responses>=0.25.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100

[tool.pytest.ini_options]
pythonpath = ["src"]
addopts = "--cov=src --cov-fail-under=80 --cov-report=term-missing"
env = [
  "ENVIRONMENT=test",
]
